pyproject.toml
[tool.poetry]
name = "lets-make-a-meal-game"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.1.2"
pusher = "^3.2.0"
bugsnag = "^4.2.1"
deepnote = "^0.2.3"
statistics = "^1.0.3"
matplotlib = "^3.6.0"
random2 = "^1.0.1"
pipedream = "^0.3.0"
replit-play = "^0.0.23"
sentry-sdk = "^1.9.8"
Everything-Tkinter = "^0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
